+++ packages/espressocash_app/lib/features/incoming_link_payments/widgets/extensions.dart
@@ -5,7 +5,6 @@ import 'package:solana/solana.dart';
 import '../../../di.dart';
 import '../../../ui/loader.dart';
 import '../../accounts/models/account.dart';
-import '../models/incoming_link_payment.dart';
 import '../services/ilp_service.dart';
 
 extension BuildContextExt on BuildContext {
@@ -22,17 +21,6 @@ extension BuildContextExt on BuildContext {
         await span.finish();
 
         return payment.id;
-      });
-
-  Future<void> retryILP(IncomingLinkPayment payment) =>
-      runWithLoader(this, () async {
-        final span = _start(function: 'retryILP');
-
-        await sl<ILPService>().retry(
-          payment,
-          account: sl<MyAccount>().wallet,
-        );
-        await span.finish();
       });
 }