+++ packages/espressocash_app/lib/features/ramp/partners/coinflow/widgets/launch.dart
@@ -115,6 +115,7 @@ extension BuildContextExt on BuildContext {
               cryptoCurrency: currency,
             ),
             partner: RampPartner.coinflow,
+            countryCode: profile.country.code,
           )
               .then((order) {
             switch (order) {