+++ packages/espressocash_app/lib/features/ramp/screens/off_ramp_order_screen.dart
@@ -128,8 +128,9 @@ class OffRampOrderScreenContent extends StatelessWidget {
         context.l10n.offRampWithdrawOngoing(
           totalAmount.format(locale),
         ),
-      OffRampOrderStatus.waitingForPartner =>
-        context.l10n.offRampWaitingForPartner,
+      OffRampOrderStatus.waitingForPartner => isMoneygramOrder
+          ? context.l10n.offRampWithdrawalInProgress
+          : context.l10n.offRampWaitingForPartner,
       OffRampOrderStatus.depositTxConfirmError ||
       OffRampOrderStatus.depositError =>
         context.l10n.offRampDepositError,
@@ -196,42 +197,40 @@ class OffRampOrderScreenContent extends StatelessWidget {
       ),
     ];
 
-    return CpTheme(
+    return StatusScreen(
       theme: theme,
-      child: StatusScreen(
-        title: context.l10n.offRampWithdrawTitle.toUpperCase(),
-        statusType: order.status.toStatusType(),
-        statusTitle: statusTitle?.let(Text.new),
-        statusContent: Column(
-          children: [
-            Text(statusContent),
-            if (order.status.isWaitingForBridge) ...bridgeSubtitleContent,
-          ],
-        ),
-        content: CpContentPadding(
-          child: SingleChildScrollView(
-            child: Column(
-              children: [
-                _Timeline(
-                  order: order,
-                  amount: totalAmount,
-                  receiveAmount: receiveAmount,
-                ),
-                if (isMoneygramOrder) _MgAdditionalInfo(order: order),
-                PartnerOrderIdWidget(orderId: order.partnerOrderId),
-                if (primaryButton != null) ...[
-                  const SizedBox(height: 12),
-                  primaryButton,
-                ],
-                Visibility(
-                  visible: showCancelButton,
-                  maintainSize: true,
-                  maintainAnimation: true,
-                  maintainState: true,
-                  child: _CancelButton(handleCanceled: handleCanceled),
-                ),
+      title: context.l10n.offRampWithdrawTitle.toUpperCase(),
+      statusType: order.status.toStatusType(),
+      statusTitle: statusTitle?.let(Text.new),
+      statusContent: Column(
+        children: [
+          Text(statusContent),
+          if (order.status.isWaitingForBridge) ...bridgeSubtitleContent,
+        ],
+      ),
+      content: CpContentPadding(
+        child: SingleChildScrollView(
+          child: Column(
+            children: [
+              _Timeline(
+                order: order,
+                amount: totalAmount,
+                receiveAmount: receiveAmount,
+              ),
+              if (isMoneygramOrder) _MgAdditionalInfo(order: order),
+              PartnerOrderIdWidget(orderId: order.partnerOrderId),
+              if (primaryButton != null) ...[
+                const SizedBox(height: 12),
+                primaryButton,
               ],
-            ),
+              Visibility(
+                visible: showCancelButton,
+                maintainSize: true,
+                maintainAnimation: true,
+                maintainState: true,
+                child: _CancelButton(handleCanceled: handleCanceled),
+              ),
+            ],
           ),
         ),
       ),
@@ -385,8 +384,11 @@ class _Timeline extends StatelessWidget {
   Widget build(BuildContext context) {
     final isMoneygramOrder = order.partner == RampPartner.moneygram;
     final CpTimelineStatus timelineStatus = order.status.toTimelineStatus();
-    final animated = timelineStatus == CpTimelineStatus.inProgress &&
-        order.status != OffRampOrderStatus.ready;
+    final animatedForMoneygram = (order.status != OffRampOrderStatus.ready &&
+            order.status != OffRampOrderStatus.waitingForPartner) ||
+        !isMoneygramOrder;
+    final animated =
+        timelineStatus == CpTimelineStatus.inProgress && animatedForMoneygram;
 
     final int activeItem = isMoneygramOrder
         ? order.status.toActiveItemForMoneygram()
@@ -400,12 +402,16 @@ class _Timeline extends StatelessWidget {
 
     final bridgingToStellar = CpTimelineItem(
       title: context.l10n.bridgingText,
-      trailing: order.bridgeAmount?.let(
-        (e) => e.isZero ? null : e.format(context.locale, maxDecimals: 2),
-      ),
     );
     final amountSent = CpTimelineItem(
-      title: context.l10n.offRampWithdrawSent,
+      title: isMoneygramOrder
+          ? context.l10n.moneygramCashAvailable
+          : context.l10n.offRampWithdrawSent,
+      trailing: isMoneygramOrder
+          ? order.bridgeAmount?.let(
+              (e) => e.isZero ? null : e.format(context.locale, maxDecimals: 2),
+            )
+          : null,
     );
     final paymentSuccess = CpTimelineItem(
       title: context.l10n.offRampWithdrawReceived,
@@ -414,10 +420,18 @@ class _Timeline extends StatelessWidget {
     );
     final paymentCanceled = CpTimelineItem(
       title: context.l10n.offRampWithdrawCancelledTitle,
+      trailing: isMoneygramOrder
+          ? order.refundAmount?.let(
+              (e) => e.isZero ? null : e.format(context.locale, maxDecimals: 2),
+            )
+          : null,
       subtitle: order.resolved?.let((t) => context.formatDate(t)),
     );
-    const refunding = CpTimelineItem(
+    final refunding = CpTimelineItem(
       title: 'Refunding USDC',
+      trailing: order.bridgeAmount?.let(
+        (e) => e.isZero ? null : e.format(context.locale, maxDecimals: 2),
+      ),
     );
 
     final normalItems = [
@@ -537,12 +551,10 @@ extension on OffRampOrderStatus {
         OffRampOrderStatus.depositTxReady ||
         OffRampOrderStatus.waitingForRefundBridge ||
         OffRampOrderStatus.sendingDepositTx ||
+        OffRampOrderStatus.waitingForPartner ||
         OffRampOrderStatus.refunded =>
           2,
-        OffRampOrderStatus.waitingForPartner ||
-        OffRampOrderStatus.failure ||
-        OffRampOrderStatus.completed =>
-          3,
+        OffRampOrderStatus.failure || OffRampOrderStatus.completed => 3,
       };
 
   bool get isRefunding =>
@@ -551,6 +563,7 @@ extension on OffRampOrderStatus {
       this == OffRampOrderStatus.waitingForRefundBridge;
 
   bool get isWaitingForBridge =>
+      this == OffRampOrderStatus.preProcessing ||
       this == OffRampOrderStatus.waitingForRefundBridge ||
       this == OffRampOrderStatus.postProcessing;