+++ packages/espressocash_app/lib/features/ramp/services/off_ramp_order_service.dart
@@ -17,6 +17,7 @@ import '../../../data/db/db.dart';
 import '../../../di.dart';
 import '../../accounts/auth_scope.dart';
 import '../../accounts/models/ec_wallet.dart';
+import '../../analytics/analytics_manager.dart';
 import '../../currency/models/amount.dart';
 import '../../currency/models/currency.dart';
 import '../../ramp_partner/models/ramp_partner.dart';
@@ -24,6 +25,7 @@ import '../../tokens/token_list.dart';
 import '../../transactions/models/tx_results.dart';
 import '../../transactions/services/resign_tx.dart';
 import '../../transactions/services/tx_sender.dart';
+import '../models/ramp_type.dart';
 import '../models/ramp_watcher.dart';
 import '../partners/coinflow/services/coinflow_off_ramp_order_watcher.dart';
 import '../partners/kado/services/kado_off_ramp_order_watcher.dart';
@@ -46,6 +48,7 @@ typedef OffRampOrder = ({
   String? authToken,
   String? referenceNumber,
   CryptoAmount? bridgeAmount,
+  CryptoAmount? refundAmount,
 });
 
 @Singleton(scope: authScope)
@@ -56,6 +59,7 @@ class OffRampOrderService implements Disposable {
     this._sender,
     this._db,
     this._tokens,
+    this._analytics,
   );
 
   final Map<String, StreamSubscription<void>> _subscriptions = {};
@@ -66,6 +70,7 @@ class OffRampOrderService implements Disposable {
   final TxSender _sender;
   final MyDatabase _db;
   final TokenList _tokens;
+  final AnalyticsManager _analytics;
 
   @PostConstruct(preResolve: true)
   Future<void> init() async {
@@ -152,6 +157,9 @@ class OffRampOrderService implements Disposable {
           .maybeWhere((it) => it.isNotEmpty)
           ?.let(Ed25519HDPublicKey.fromBase58);
 
+      final refundAmount = row.refundAmount
+          ?.let((it) => CryptoAmount(value: it, cryptoCurrency: Currency.usdc));
+
       return (
         id: row.id,
         created: row.created,
@@ -169,6 +177,7 @@ class OffRampOrderService implements Disposable {
         authToken: row.authToken,
         referenceNumber: row.referenceNumber,
         bridgeAmount: bridgeAmount,
+        refundAmount: refundAmount,
       );
     });
   }
@@ -265,6 +274,7 @@ class OffRampOrderService implements Disposable {
     (SignedTx, BigInt)? transaction,
     FiatAmount? receiveAmount,
     CryptoAmount? fee,
+    required String countryCode,
   }) =>
       tryEitherAsync((_) async {
         {
@@ -293,6 +303,14 @@ class OffRampOrderService implements Disposable {
           _subscribe(order.id);
           await _watch(order.id);
 
+          _analytics.rampInitiated(
+            partner: partner,
+            rampType: RampType.offRamp.name,
+            amount: amount.value.toString(),
+            countryCode: countryCode,
+            id: order.id,
+          );
+
           return order.id;
         }
       });
@@ -304,6 +322,7 @@ class OffRampOrderService implements Disposable {
     required RampPartner partner,
     required BigInt slot,
     FiatAmount? receiveAmount,
+    required String countryCode,
   }) =>
       tryEitherAsync((bind) async {
         {
@@ -316,6 +335,7 @@ class OffRampOrderService implements Disposable {
             depositAddress: '',
             receiveAmount: receiveAmount,
             transaction: (signed, slot),
+            countryCode: countryCode,
           ).letAsync(bind);
         }
       });