+++ packages/espressocash_app/lib/features/ramp/widgets/ramp_buttons.dart
@@ -10,6 +10,7 @@ import '../../../gen/assets.gen.dart';
 import '../../../ui/button.dart';
 import '../../../ui/icon_button.dart';
 import '../../accounts/models/account.dart';
+import '../../analytics/analytics_manager.dart';
 import '../../country_picker/models/country.dart';
 import '../../feature_flags/services/feature_flags_manager.dart';
 import '../../profile/data/profile_repository.dart';
@@ -245,10 +246,13 @@ extension RampBuildContextExt on BuildContext {
       case RampPartner.scalex:
         launchScalexOnRamp(profile: profile, address: address);
       case RampPartner.moneygram:
-        launchMoneygramOnRamp();
+        launchMoneygramOnRamp(profile: profile);
       case RampPartner.coinflow:
         throw UnimplementedError('Not implemented for $partner');
     }
+
+    sl<AnalyticsManager>()
+        .rampOpened(partner: partner, rampType: RampType.onRamp.name);
   }
 
   void _launchOffRampPartner(
@@ -264,11 +268,14 @@ extension RampBuildContextExt on BuildContext {
       case RampPartner.scalex:
         launchScalexOffRamp(profile: profile, address: address);
       case RampPartner.moneygram:
-        launchMoneygramOffRamp();
+        launchMoneygramOffRamp(profile: profile);
       case RampPartner.rampNetwork:
       case RampPartner.guardarian:
         throw UnimplementedError('Not implemented for $partner');
     }
+
+    sl<AnalyticsManager>()
+        .rampOpened(partner: partner, rampType: RampType.offRamp.name);
   }
 }
 
@@ -294,7 +301,7 @@ IList<RampPartner> _getOnRampPartners(String countryCode) {
   final isMoneygramEnabled =
       sl<FeatureFlagsManager>().isMoneygramAccessEnabled();
 
-  if (isMoneygramEnabled && _moneygramCountries.contains(countryCode)) {
+  if (isMoneygramEnabled && _moneygramOnRampCountries.contains(countryCode)) {
     partners.add(RampPartner.moneygram);
   }
 
@@ -315,7 +322,7 @@ IList<RampPartner> _getOffRampPartners(String countryCode) {
   final isMoneygramEnabled =
       sl<FeatureFlagsManager>().isMoneygramAccessEnabled();
 
-  if (isMoneygramEnabled && _moneygramCountries.contains(countryCode)) {
+  if (isMoneygramEnabled && _moneygramOffRampCountries.contains(countryCode)) {
     partners.add(RampPartner.moneygram);
   }
 
@@ -338,4 +345,5 @@ const _coinflowCountries = {
 
 const _scalexCountries = {'NG'};
 
-const _moneygramCountries = {'US'};
+const _moneygramOnRampCountries = {'US'};
+const _moneygramOffRampCountries = {'US', 'PT'};